requirements_local.txt
asgiref==3.8.1
attrs==25.1.0
Django==5.1.6
django-cors-headers==4.7.0
django-filter==24.3
django-leaflet==0.31.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.9
GDAL==3.9.3
geographiclib==2.0
geojson==3.2.0
geopy==2.4.1
gis==0.2.1
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==2.2.3
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
setuptools==75.8.2
shapely==2.0.7
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
